refactorings and minor bug fix
