+++ jubako/src/main/java/com/justeat/jubako/JubakoAdapter.kt
@@ -1,4 +1,3 @@
-/*
 package com.justeat.jubako
 
 import android.os.Handler
@@ -6,86 +5,44 @@ import android.view.ViewGroup
 import androidx.lifecycle.LifecycleOwner
 import androidx.recyclerview.widget.RecyclerView
 import com.justeat.jubako.JubakoViewHolder.Event
-import com.justeat.jubako.util.IJubakoScreenFiller
+import com.justeat.jubako.adapters.DefaultProgressViewHolder
+import com.justeat.jubako.adapters.ProgressAdapter
+import com.justeat.jubako.data.PaginatedDataState
+import com.justeat.jubako.extensions.CreateViewHolderDelegate
 import com.justeat.jubako.util.JubakoScreenFiller
 import com.justeat.jubako.widgets.JubakoRecyclerView
 
 open class JubakoAdapter(
     private val lifecycleOwner: LifecycleOwner,
     private val data: Jubako.Data,
-    private val loadingStrategy: ContentLoadingStrategy = PaginatedContentLoadingStrategy()
-) : RecyclerView.Adapter<JubakoViewHolder<Any>>() {
-
+    private val loadingStrategy: ContentLoadingStrategy = PaginatedContentLoadingStrategy(),
+    override val progressViewHolder: CreateViewHolderDelegate<RecyclerView.ViewHolder> = { inflater, parent, _ ->
+        DefaultProgressViewHolder(inflater, parent)
+    }
+) : ProgressAdapter<Any, JubakoViewHolder<Any>>(
+    Jubako.logger, progressViewHolder, JubakoScreenFiller.Orientation.VERTICAL
+) {
+    private val handler = Handler()
     var onInitialFill: () -> Unit = {}
     var onViewHolderBound: (contentViewHolder: JubakoViewHolder<Any>) -> Unit = {}
 
-    private val handler = Handler()
-    var logger = Jubako.logger
     var hasMore: Boolean = true
-
-    var screenFiller = IJubakoScreenFiller.NOOP
+    var state = PaginatedDataState<ContentDescription<Any>>(listOf(), listOf(), true, null, false)
 
     interface HolderFactory<T> {
         fun createViewHolder(parent: ViewGroup): JubakoViewHolder<T>
     }
 
-    override fun onAttachedToRecyclerView(recyclerView: RecyclerView) {
-        logger.log(TAG, "On Attach", "total rows: ${data.source.size}")
-        (recyclerView as JubakoRecyclerView).onDrawComplete = {}
-        setupLoadMoreScrollTrigger(recyclerView)
-        initialFillOnLayoutChanged(recyclerView)
-        listenForContentChanges()
-    }
-
-    private fun listenForContentChanges() {
-        //
-        // Listens for changes in ContentDescriptionCollection and relays
-        // then to the adapter via the PaginatedDataAdapter notify methods
-        //
-        data.destination.listener =
-            ContentAdapterContentDescriptionCollectionListener(this)
-    }
-
-    private fun initialFillOnLayoutChanged(recyclerView: RecyclerView) {
-        logger.log(TAG, "Initial Fill Down", "begin...")
-        screenFiller = JubakoScreenFiller(
-            orientation = JubakoScreenFiller.Orientation.VERTICAL,
-            logger = logger,
-            log = true,
-            hasMore = { hasMore },
-            loadMore = this::load,
-            onFilled = {
-                onInitialFill()
-            }
-        )
-        screenFiller.attach(recyclerView)
+    override fun init(recyclerView: RecyclerView) {
         load()
+        (recyclerView as JubakoRecyclerView).onDrawComplete = {} // TODO why?
+        data.destination.listener = ContentAdapterContentDescriptionCollectionListener(this)
     }
 
-    private var inLoadingError = false
-
-    private fun load() {
-        if (!inLoadingError) {
-            loadingStrategy.load(lifecycleOwner, data,
-                onLoaded = {
-                    hasMore = it
-                    false
-                },
-                onError = {
-                    inLoadingError = true
-                })
-        }
-    }
-
-    override fun getItemCount(): Int {
-        return data.numItemsLoaded()
-    }
-
-    override fun onCreateViewHolder(viewGroup: ViewGroup, viewType: Int): JubakoViewHolder<Any> {
-
+    override fun onCreateViewHolderItem(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         val factory = data.viewHolderFactories[viewType]
 
-        val holder = factory.createViewHolder(viewGroup)
+        val holder = factory.createViewHolder(parent)
         holder.logger = logger
 
         logger.log(
@@ -96,17 +53,20 @@ open class JubakoAdapter(
         return holder
     }
 
-    override fun onBindViewHolder(holder: JubakoViewHolder<Any>, position: Int) {
-        logger.log(TAG, "Bind ViewHolder")
+    override fun getItemCountActual(): Int = data.numItemsLoaded()
+    override fun getCurrentState(): PaginatedDataState<*> = state
+    override fun loadMore() = load()
+    override fun hasMoreToLoad(): Boolean {
+        return hasMore
+    }
 
-        val item = data.getItem(position)
+    override fun getItem(position: Int): Any = data.getItem(position)
 
-        holder.onClickDelegate = { postViewHolderEvent(Event.Click(item.id, it)) }
-
-        attachReloader(item, holder)
-
-        bindWhenNewOrDefault(item, holder)
-
+    override fun bindItemToHolder(holder: JubakoViewHolder<Any>, item: Any) {
+        val description = (item as ContentDescription<Any>)
+        holder.onClickDelegate = { postViewHolderEvent(Event.Click(description.id, it)) }
+        attachReloader(description, holder)
+        bindWhenNewOrDefault(description, holder)
         onViewHolderBound(holder)
     }
 
@@ -117,6 +77,16 @@ open class JubakoAdapter(
             holder.bind(data)
             holder.data = data
         }
+    }
+
+    override fun onFilled() = onInitialFill()
+
+    private fun load() = loadingStrategy.load(
+        lifecycleOwner, data
+    ) { newState, more ->
+        hasMore = more
+        onStateChanged(newState, state)
+        state = newState
     }
 
     private fun attachReloader(item: ContentDescription<Any>, holder: JubakoViewHolder<Any>) {
@@ -137,33 +107,8 @@ open class JubakoAdapter(
         }
     }
 
-    override fun getItemViewType(position: Int): Int {
+    override fun getItemViewTypeActual(position: Int): Int {
         return data.getItemViewType(position)
-    }
-
-    private var scrollListener: RecyclerView.OnScrollListener? = null
-
-    private fun setupLoadMoreScrollTrigger(recyclerView: RecyclerView) {
-        if (scrollListener != null) {
-            recyclerView.removeOnScrollListener(scrollListener!!)
-            scrollListener = null
-        }
-
-        scrollListener = object : RecyclerView.OnScrollListener() {
-            override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
-                val offset = recyclerView.computeVerticalScrollOffset()
-                val range = recyclerView.computeVerticalScrollRange() - recyclerView.computeVerticalScrollExtent()
-                if (range != 0 && offset > range * 0.8f) {
-                    logger.log(TAG, "Scroll Trigger Load")
-                    loadingStrategy.load(lifecycleOwner, data, onLoaded = {
-                        logger.log(TAG, "Scroll Trigger Load Complete")
-                        false
-                    })
-                }
-            }
-        }
-
-        recyclerView.addOnScrollListener(scrollListener!!)
     }
 
     fun reload(contentDescriptionId: String, payload: Any? = null) {
@@ -178,16 +123,10 @@ open class JubakoAdapter(
         }
     }
 
-    override fun getItemId(position: Int): Long {
-        return data.getItemId(position)
-    }
-
-    private fun postViewHolderEvent(event: Event) {
-        onViewHolderEvent(event)
-    }
+    override fun getItemId(position: Int): Long = data.getItemId(position)
+    private fun postViewHolderEvent(event: Event) = onViewHolderEvent(event)
 
     var onViewHolderEvent: (event: Event) -> Unit = { }
 }
 
 private val TAG = JubakoAdapter::class.java.simpleName
-*/
