+++ jubako/src/main/java/com/justeat/jubako/adapters/PaginatedDataAdapter.kt
@@ -8,18 +8,19 @@ import androidx.recyclerview.widget.RecyclerView
 import com.justeat.jubako.Jubako
 import com.justeat.jubako.data.PaginatedDataState
 import com.justeat.jubako.data.PaginatedLiveData
+import com.justeat.jubako.extensions.CreateViewHolderDelegate
 import com.justeat.jubako.util.JubakoScreenFiller
 
 @Suppress("UNCHECKED_CAST")
 class PaginatedDataAdapter<DATA, ITEM_DATA, ITEM_HOLDER : RecyclerView.ViewHolder>(
     override var logger: Jubako.Logger,
-    var itemViewHolder: (layoutInflater: LayoutInflater, parent: ViewGroup, viewType: Int) -> ITEM_HOLDER,
+    var itemViewHolder: CreateViewHolderDelegate<ITEM_HOLDER>,
     var itemBinder: (holder: ITEM_HOLDER, data: ITEM_DATA?) -> Unit = { _, _ -> },
     var lifecycleOwner: LifecycleOwner,
     var itemData: (data: DATA, position: Int) -> ITEM_DATA,
     var itemCount: (data: DATA) -> Int,
     var paginatedLiveData: PaginatedLiveData<*>,
-    override val progressViewHolder: (parent: ViewGroup) -> RecyclerView.ViewHolder,
+    override val progressViewHolder: CreateViewHolderDelegate<RecyclerView.ViewHolder>,
     override var orientation: JubakoScreenFiller.Orientation
 ) : ProgressAdapter<ITEM_DATA, ITEM_HOLDER>(
     logger, progressViewHolder, orientation
