+++ jubako/src/main/java/com/justeat/jubako/adapters/StaticDataAdapter.kt
@@ -4,12 +4,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.justeat.jubako.Jubako
+import com.justeat.jubako.extensions.CreateViewHolderDelegate
 
 @Suppress("UNCHECKED_CAST")
 class StaticDataAdapter<DATA, ITEM_DATA, ITEM_HOLDER : RecyclerView.ViewHolder>(
     private val data: DATA,
     var logger: Jubako.Logger,
-    var itemViewHolder: (layoutInflater: LayoutInflater, parent: ViewGroup, viewType: Int) -> ITEM_HOLDER,
+    var itemViewHolder: CreateViewHolderDelegate<ITEM_HOLDER>,
     var itemBinder: (holder: ITEM_HOLDER, data: ITEM_DATA?) -> Unit = { _, _ -> },
     var itemData: (data: DATA, position: Int) -> ITEM_DATA,
     var itemCount: (data: DATA) -> Int
