+++ jubako/src/main/java/com/justeat/jubako/extensions/JubakoAssembleExtensions.kt
@@ -1,9 +1,15 @@
 package com.justeat.jubako.extensions
 
 import androidx.lifecycle.LiveData
-import com.justeat.jubako.*
+import com.justeat.jubako.ContentDescription
+import com.justeat.jubako.ContentDescriptionProvider
+import com.justeat.jubako.JubakoAdapter
+import com.justeat.jubako.JubakoAssembler
+import com.justeat.jubako.JubakoViewHolder
 import com.justeat.jubako.data.EmptyLiveData
-import java.util.*
+import com.justeat.jubako.descriptionProvider
+import com.justeat.jubako.viewHolderFactory
+import java.util.UUID
 
 typealias ListReceiver = MutableList<ContentDescriptionProvider<Any>>.() -> Unit
 typealias JubakoMutableList = MutableList<ContentDescriptionProvider<Any>>
@@ -32,7 +38,7 @@ fun <T> JubakoMutableList.addHolder(delegate: () -> JubakoViewHolder<T>) {
     })
 }
 
-fun <T> MutableList<ContentDescriptionProvider<Any>>.addDescription(
+fun <T> JubakoMutableList.addDescription(
     viewHolderFactory: JubakoAdapter.HolderFactory<T>,
     data: LiveData<T> = EmptyLiveData(),
     onReload: (ContentDescription<T>.(payload: Any?) -> Unit) = { },
