+++ jubako/src/main/res/layout/default_progress.xml
@@ -1,17 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="24dp"
-    android:layout_height="24dp">
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content" android:layout_gravity="center">
 
-    <View
+    <ProgressBar
         android:id="@+id/progress"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:background="#0000ff" />
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:indeterminate="true" />
 
-    <View
+    <Button
         android:id="@+id/button_retry"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:background="#ff0000" />
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:drawableLeft="@android:drawable/progress_horizontal"
+        android:text="Retry" />
 </FrameLayout>
